feat: 저장소 목록 화면에 데이터 연동
