+++ README.md
@@ -5,9 +5,9 @@
 - [x] 저장소 목록을 보여주는 화면 구현
     - [x] 상단바 추가
     - [x] 저장소 목록 추가
-- [] 저장소 목록 데이터를 관리하는 ViewModel 구현
-- [] 저장소 목록 화면에 데이터 연동
-- [] 저장소 목록 화면 리팩터링
+- [x] 저장소 목록 데이터를 관리하는 ViewModel 구현
+- [x] 저장소 목록 화면에 데이터 연동
+- [x] 저장소 목록 화면 리팩터링
 
 ## Step2 진행 중 의식의 흐름
 -
